rename projethome to projectgrid
