+++ src/main/kotlin/org/wycliffeassociates/otter/jvm/app/ui/projectgrid/view/ProjectGridStyles.kt
@@ -1,0 +1,59 @@
+package org.wycliffeassociates.otter.jvm.app.ui.projecthome.view
+
+import javafx.geometry.Pos
+import javafx.scene.Cursor
+import javafx.scene.text.FontWeight
+import javafx.stage.Screen
+import org.wycliffeassociates.otter.jvm.app.theme.AppTheme
+import tornadofx.*
+
+class ProjectHomeStyles : Stylesheet() {
+    companion object {
+        val homeAnchorPane by cssclass()
+        val projectsFlowPane by cssclass()
+        val noProjectsLabel by cssclass()
+        val tryCreatingLabel by cssclass()
+        val addProjectButton by cssclass()
+    }
+
+    init {
+        homeAnchorPane {
+            prefWidth = Screen.getPrimary().visualBounds.width.px - 50.0
+            prefHeight = Screen.getPrimary().visualBounds.height.px - 50.0
+        }
+
+        projectsFlowPane {
+            vgap = 32.px
+            hgap = 24.px
+            alignment = Pos.TOP_LEFT
+            // Add larger padding on bottom to keep FAB from blocking last row cards
+            padding = box(10.px, 20.px, 95.px, 20.px)
+        }
+
+        noProjectsLabel {
+            fontSize = 30.px
+            fontWeight = FontWeight.BOLD
+            textFill = AppTheme.colors.defaultText
+        }
+
+        tryCreatingLabel {
+            fontSize = 20.px
+            textFill = AppTheme.colors.defaultText
+        }
+
+        addProjectButton {
+            unsafe("-jfx-button-type", raw("RAISED"))
+            backgroundRadius += box(25.px)
+            borderRadius += box(25.px)
+            backgroundColor += AppTheme.colors.white
+            minHeight = 50.px
+            minWidth = 50.px
+            maxHeight = 50.px
+            maxWidth = 50.px
+            cursor = Cursor.HAND
+            child("*") {
+                fill = AppTheme.colors.appRed
+            }
+        }
+    }
+}
